pyproject.toml
[tool.poetry]
name = "shogi-camera-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
black = "^21.11b1"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.910"
tox = "^3.24.4"
opencv-python = "^4.5.4"
scipy = "^1.7.3"
firebase-admin = "^5.1.0"
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79